% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_C_M}
\alias{compute_C_M}
\title{Compute C(M) - Cost of Refinancing Normalized for Tax Rate}
\usage{
compute_C_M(kappa, tau)
}
\arguments{
\item{kappa}{The real, tax-adjusted cost of refinance which
includes points and any other explicit or implicit transaction
costs.  Kappa(M) is the net present value (NPV) of these costs,
netting out all allowable tax deductions generated by future
deductions of amortized refinancing points.}

\item{tau}{The marginal tax rate.}
}
\value{
C(M): The cost of refinancing normalized for the tax rate, tau.
}
\description{
Calculates the cost of refinancing normalized for the tax rate,
based on the real, tax-adjusted cost of refinance.
}
\examples{
# compute_C_M(2000, 0.25)

}
\references{
Agarwal et al. (2013), p. 595 (p. 5 of pdf).
}
