% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_kappa_M_itemize_mtg_pymts}
\alias{compute_kappa_M_itemize_mtg_pymts}
\title{Compute kappa(M) for Itemized Mortgage Payments}
\usage{
compute_kappa_M_itemize_mtg_pymts(F, f, M, tau, mu, rho, pi, N)
}
\arguments{
\item{F}{The fixed cost of refinancing.}

\item{f}{\code{100 * f} is the number of points (so, specify the points as a decimal).}

\item{M}{The real value of remaining principal.}

\item{tau}{The marginal tax rate.}

\item{mu}{Probability of moving each year (as a decimal); Agarwal et al. assume 0.1
or an expected duration of staying in the house of 10 years.}

\item{rho}{The real discount rate.}

\item{pi}{The inflation rate.}

\item{N}{The number of years for the new mortgage after refinancing.}
}
\value{
Kappa(M) -- the real, tax-adjusted cost of refinance. These
costs include points and any other explicit or implicit
transaction costs (such as lawyers fees, mortgage insurance,
personal time). Kappa(M) is the net present value (NPV) of these
costs, netting out all allowable tax deductions generated by
future deductions of amortized refinancing points.
}
\description{
Calculates the real, tax-adjusted cost of refinancing a mortgage, considering various financial parameters
and costs involved in the process. Implements the formula described in Agarwal et al. (2013), specifically equation A3 in Appendix A.
}
\examples{
compute_kappa_M_itemize_mtg_pymts(2000, 0.01, 100000, 0.25, 0.1, 0.03, 0.02, 30)

}
\references{
Agarwal et al. (2013), p. 595 (p. 5 of pdf), Equation A3
}
