% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_kappa_M_no_itemizing_of_mtg_pymts}
\alias{compute_kappa_M_no_itemizing_of_mtg_pymts}
\title{Compute kappa(M) Without Itemizing Mortgage Payments}
\usage{
compute_kappa_M_no_itemizing_of_mtg_pymts(f, M, F)
}
\arguments{
\item{f}{Points where \code{100 * f} is the number of points (so \code{f} is
the points in decimal form)}

\item{M}{The real value of the remaining principal.}

\item{F}{The fixed transaction costs of refinancing (e.g.,lawyers fees, mortgage
insurance, personal time).}
}
\value{
Kappa(M) -- the real, tax-adjusted cost of refinance. These
costs include points and any other explicit or implicit
transaction costs. Kappa(M) is the net present value (NPV) of
these costs, netting out all allowable tax deductions generated
by future deductions of amortized refinancing points.
}
\description{
Calculates the real, tax-adjusted cost of refinancing a mortgage
without itemizing deductions, considering points, the remaining
principal, and fixed transaction costs. Implements the calculation
as per Agarwal et al. (2013), referenced in equation A3 in Appendix
A.
}
\examples{
compute_kappa_M_no_itemizing_of_mtg_pymts(0.01, 100000, 2000)

}
\references{
Agarwal et al. (2013), Equation A3, p. 595 (p. 5 of pdf).
}
