% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_lambda}
\alias{compute_lambda}
\title{Compute lambda}
\usage{
compute_lambda(mu, i_0, gamma_uppercase, pi)
}
\arguments{
\item{mu}{Probability of moving each year (as a decimal); Agarwal
et al. assume 0.1 or an expected duration of staying in the house
of 10 years.}

\item{i_0}{The nominal interest rate of the mortgage.}

\item{gamma_uppercase}{The remaining life of the mortgage in years.}

\item{pi}{The inflation rate.}
}
\value{
The expected real rate of exogenous mortgage repayment.
}
\description{
Calculates the expected real rate of exogenous mortgage repayment
based on various financial parameters.  This function implements
the formula described on page 603 (p. 13 of pdf) from Agarwal et
al. (2013).
}
\examples{
lambda_agarwal_et_al_p_603_example <- compute_lambda(
 mu = 0.1, i_0 = 0.06, gamma_uppercase = 25, pi = 0.03
) |>
 round(3)
stopifnot(lambda_agarwal_et_al_p_603_example == 0.147)

}
\references{
Agarwal et al., 2013, p. 603 (p. 13 of pdf)
}
