% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_optimal_refi_rates.R
\name{compute_optimal_refi_rates}
\alias{compute_optimal_refi_rates}
\title{Compute Optimal Refinance Rates}
\usage{
compute_optimal_refi_rates(
  rho,
  sigma,
  mu,
  i_0,
  gamma_uppercase,
  pi,
  tau,
  M,
  F,
  f,
  N,
  kappa = NULL,
  itemize_mtg_pymts = FALSE
)
}
\arguments{
\item{rho}{The real discount rate, as a decimal, which equals the
nominal interest rate minus the inflation rate.  See Agarwal et
al. (2013) p. 606 (p. 16 of the pdf) and the Fisher equation for
details.  Fisher equation:
https://en.wikipedia.org/wiki/Fisher_equation}

\item{sigma}{The annualized standard deviation of the mortgage
interest rate.  See footnote 15 in Agarwal et al. (2013) for
details.}

\item{mu}{Probability of moving each year (as a decimal); Agarwal
et al. assume 0.1 or an expected duration of staying in the house
of 10 years.}

\item{i_0}{The nominal interest rate of the mortgage.}

\item{gamma_uppercase}{The remaining life of the mortgage in years.}

\item{pi}{Inflation rate as a decimal.}

\item{tau}{The marginal tax rate as a decimal.}

\item{M}{The real value of remaining principal.}

\item{F}{Fixed costs of refinancing.}

\item{f}{Fraction of the mortgage amount representing the variable
costs of refinancing. \code{100 * f} is the number of points (so,
specify the points as a decimal).}

\item{N}{The number of years for the new mortgage after
refinancing.}

\item{kappa}{The real, tax-adjusted cost of refinance. These costs
include points and any other explicit or implicit transaction
costs (such as lawyers fees, mortgage insurance, personal
time). kappa is the net present value (NPV) of these costs,
netting out all allowable tax deductions generated by future
deductions of amortized refinancing points. The default value of
kappa is \code{NULL}, where the function will calculate kappa based on
other parameters. If the user supplies \code{kappa}, that will be used
and \code{kappa} will not be calculated.}

\item{itemize_mtg_pymts}{Logical; if TRUE, mortgage payments are
itemized in the tax-adjusted cost of refinance. if FALSE,
mortgage payments are not a tax deduction for the tax-adjusted
cost of refinance. If the user supplies \code{kappa}, this parameter
will not affect output.}
}
\value{
A data frame with the following columns:
\itemize{
\item x_star_optimal_refi_differential: The optimal refinance rate differential from Agarwal et al. (2013).
\item max_refi_rate_using_x_star: The maximum refinance rate using x_star rule.
\item x_star_sqrt_rule: The square root rule for optimal refinance rate differential from Agarwal et al. (2013).
\item max_refi_rate_using_x_star_sqrt_rule: The maximum refinance rate using the square root rule.
\item x_pv_rule: The present value rule for the refinance rate differential.
\item max_refi_rate_using_x_pv_rule: The maximum refinance rate using the present value rule.
\item lambda: The expected real rate of exogenous mortgage repayment.
\item kappa: The real, tax-adjusted cost of refinancing a mortgage.
\item C_M: The cost of refinancing normalized for the tax rate, tau.
\item psi: psi value as described in the referenced source. \href{https://chat.openai.com/share/90c95875-8940-4979-aaf2-bc0acff5282b}{ChatGPT explanation of psi}.
\item phi: Phi value as described in the referenced source.
}
}
\description{
This function calculates the optimal refinance rates based on various financial parameters and decision thresholds. It incorporates tax adjustments, mortgage payment itemization, and different rules to determine the optimal refinance differential.
}
\examples{
# Example, using the calibration in Agarwal table 1 for a $1M mortgage:
optimal_rates <- compute_optimal_refi_rates(
  rho = 0.05, sigma = 0.0109, mu = 0.1, i_0 = 0.06,
  gamma_uppercase = 25, pi = 0.03, tau = 0.28,
  M = 1e06, F = 2000, f = 0.01, N = 30,
  itemize_mtg_pymts = TRUE
)

}
\references{
Agarwal, S., Driscoll, J. C., & Laibson, D. (2013). Optimal Mortgage Refinancing: A Closed-Form Solution.
Journal of Money, Credit and Banking, 45(4), 591-622.
\href{https://en.wikipedia.org/wiki/Fisher_equation}{Wikipedia Fisher Equation}
\href{https://chat.openai.com/share/4b0d0bb5-8572-4540-8841-24a5cacdffe6}{ChatGPT Fisher Equation}
\href{https://chat.openai.com/share/90c95875-8940-4979-aaf2-bc0acff5282b}{ChatGPT explanation of psi}
}
