% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_phi}
\alias{compute_phi}
\title{Compute phi}
\usage{
compute_phi(psi, rho, lambda, kappa, M, tau)
}
\arguments{
\item{psi}{Input from the compute_psi function.}

\item{rho}{The real discount rate, which equals the nominal
interest rate minus the inflation rate.  See Agarwal et
al. (2013) p. 606 (p. 16 of the pdf) and the Fisher equation for
details.}

\item{lambda}{The expected real rate of exogenous mortgage
repayment.}

\item{kappa}{The real, tax-adjusted cost of refinance which
includes points and any other explicit or implicit transaction
costs.  Kappa(M) is the net present value (NPV) of these costs,
netting out all allowable tax deductions generated by future
deductions of amortized refinancing points.}

\item{M}{The real value of the remaining principal.}

\item{tau}{The marginal tax rate.}
}
\value{
Phi value as described in the referenced source.
}
\description{
Calculates the phi value based on the psi value from the
compute_psi function, the real discount rate, the expected real
rate of exogenous mortgage repayment, the real, tax-adjusted cost
of refinance, the remaining principal, and the marginal tax
rate.
}
\examples{
psi_value <- compute_psi(rho = 0.05, lambda = 0.147, sigma = 0.0109)
compute_phi(psi = psi_value, rho = 0.05, lambda = 0.147, kappa = 5000,
            M = 100000, tau = 0.28)

}
\references{
Agarwal et al. (2013)
Fisher Equation: https://en.wikipedia.org/wiki/Fisher_equation
}
