% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_psi}
\alias{compute_psi}
\title{Compute psi}
\usage{
compute_psi(rho, lambda, sigma)
}
\arguments{
\item{rho}{The real discount rate, which equals the nominal
interest rate minus the inflation rate.
See Agarwal et al. (2013) p. 606 (p. 16 of the pdf) and the Fisher equation for
details.
Fisher equation: https://en.wikipedia.org/wiki/Fisher_equation}

\item{lambda}{The expected real rate of exogenous mortgage
repayment.}

\item{sigma}{The annualized standard deviation of the mortgage
interest rate.  See footnote 15 in Agarwal et al. (2013) for
details.}
}
\value{
psi value as described in the referenced source. \href{https://chat.openai.com/share/90c95875-8940-4979-aaf2-bc0acff5282b}{ChatGPT explanation of psi}
}
\description{
Calculates the psi value based on the real discount rate, the
expected real rate of exogenous mortgage repayment, and the
annualized standard deviation of the mortgage interest rate. This
function implements the formula and parameters as described in
Agarwal et al. (2013).
}
\examples{
compute_psi(rho = 0.05, lambda = 0.147, sigma = 0.0109)
}
\references{
Agarwal et al., 2013, p. 603 (p. 13 of pdf)
\href{https://en.wikipedia.org/wiki/Fisher_equation}{Wikipedia Fisher Equation}
\href{https://chat.openai.com/share/4b0d0bb5-8572-4540-8841-24a5cacdffe6}{ChatGPT Fisher Equation}
}
