% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_pv_breakeven_threshold}
\alias{compute_pv_breakeven_threshold}
\title{Compute Present Value Breakeven Threshold}
\usage{
compute_pv_breakeven_threshold(C_M, rho, lambda, M)
}
\arguments{
\item{C_M}{Cost of refinancing normalized for the tax rate, tau.}

\item{rho}{The real discount rate. Using the Fisher equation, the
real discount rate equals nominal interest rate minus the
inflation rate.}

\item{lambda}{The expected real rate of exogenous mortgage
repayment.}

\item{M}{The real value of the remaining principal.}
}
\value{
x_pv: The present value breakeven threshold for refinancing a mortgage.
}
\description{
Calculates the present value breakeven threshold for refinancing a
mortgage. This threshold is based on the cost of refinancing
normalized for the tax rate, the real discount rate, the expected
real rate of exogenous mortgage repayment, and the remaining
principal. Implements the calculation as per equation 22 in Agarwal
et al. (2013).
}
\examples{
compute_pv_breakeven_threshold(C_M = 2000, rho = 0.05, lambda = 0.05, M = 100000)

}
