% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_x_star}
\alias{compute_x_star}
\title{Compute x*}
\usage{
compute_x_star(rho, sigma, mu, i_0, gamma_uppercase, pi, tau, M, kappa)
}
\arguments{
\item{rho}{The real discount rate, which equals the nominal
interest rate minus the inflation rate.}

\item{sigma}{The annualized standard deviation of the mortgage
interest rate.}

\item{mu}{Probability of moving each year (as a decimal).}

\item{i_0}{The nominal interest rate of the mortgage.}

\item{gamma_uppercase}{The remaining life of the mortgage in years.}

\item{pi}{The inflation rate.}

\item{tau}{The marginal tax rate.}

\item{M}{The real value of the remaining principal.}

\item{kappa}{The real, tax-adjusted cost of refinance (see p. 595
p. 5 of the pdf of Agarwal et al. 2013). These costs include
points and any other explicit or implicit transaction costs (such
as lawyers fees, mortgage insurance, personal time). kappa(M)
is the net present value (NPV) of these costs, netting out all
allowable tax deductions generated by future deductions of
amortized refinancing point.}
}
\value{
x_star: The threshold for the refinance decision,
indicating that the borrower should refinance when the current
mortgage interest rate minus the original mortgage interest rate
is less than or equal to x_star.
}
\description{
Calculates the threshold for the refinance decision so that the
borrower should refinance when the current mortgage interest rate
minus the original mortgage interest rate is less than or equal to
x*.
}
\details{
The borrower should refinance when i - i_0 <= x_star,
where i represents the current mortgage interest rate and i_0
represents the original mortgage interest rate.
}
\examples{
compute_x_star(rho = 0.03, sigma = 0.1, mu = 0.1, i_0 = 0.05,
               gamma_uppercase = 30, pi = 0.02, tau = 0.25, M = 100000,
               kappa = 9904.783)

}
