% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_x_star_short}
\alias{compute_x_star_short}
\title{Compute x* (Short Version)}
\usage{
compute_x_star_short(psi, phi)
}
\arguments{
\item{psi}{Psi value, likely from the compute_psi() function.}

\item{phi}{Phi value, likely from the compute_phi() function.}
}
\value{
x_star: The threshold for the refinance decision.
}
\description{
Calculates the threshold for the refinance decision so that the
borrower should refinance when the current mortgage interest rate
minus the original mortgage interest rate is less than or equal to
x*. This function uses the psi and phi values, likely obtained from
the compute_psi() and compute_phi() functions, respectively.
}
\details{
The borrower should refinance when i - i_0 <= x_star,
where i represents the current mortgage interest rate and i_0
represents the original mortgage interest rate. The calculation
involves the main branch of the Lambert W function, specifically
gsl::lambert_W0 in the GNU Scientific Library.
}
\examples{
psi_value <- compute_psi(rho = 0.05, lambda = 0.147, sigma = 0.0109)
phi_value <- compute_phi(psi_value, rho = 0.05, lambda = 0.147,
                         kappa = 9904.783, M = 1e06, tau = 0.28)
compute_x_star_short(psi_value, phi_value)

}
